Add the edit order feature

+++ database/migrations/2022_03_26_033406_create_item_order_table.php
@@ -18,7 +18,7 @@ class CreateItemOrderTable extends Migration
             $table->foreignId('order_id')->constrained();
             $table->foreignId('item_id')->constrained();
             $table->foreignId('price_id')->constrained();
-            $table->tinyInteger('quantity');
+            $table->smallInteger('quantity');
         });
     }
